Normal maps!
Implemented normal maps to go with my new awesome arwing model and its
crazy-wicked normal map texture. Looks so friggi'n balli'n that I had to
disable the sun and enable the disco lights again just to bask in its
glory.

Also fixed misc. compile-warning messages.

+++ HUDScene.cpp
@@ -6,7 +6,7 @@
 HUDScene::HUDScene() {
 
 	glm::ivec2 w_size = Game::instance()->getWindowSize();
-	GameObject* camera = GOFactory::instance()->createGUICamera(w_size.x, w_size.y);
+	GameObject* camera = GOFactory::instance()->createGUICamera((float)w_size.x, (float)w_size.y);
 	add(camera);
 
 	useCamera((CameraComponent*)camera->getComponent(Component::CAMERA));
